feat(State): Switch Internal State Object to prototypally inherit from the State Declaration
In 0.x, we used prototypal inheritance, i.e. `internalImpl = inherit(stateDeclaration)`.
In 1.0 alphas, we switched to `extend(new State(), stateDeclaration)` to enable us to use `instanceof State`.
However, this convenience for ui-router code was inconvenient for end users so we're switching back to using the state declaration as the prototype of the internal state object.

- This enables users to create classes for their states, providing (e.g.) standardized onEnter functions, such as:
```js
var myState = new LoggingState('mystate', '/url');
```

Closes https://github.com/angular-ui/ui-router/issues/3293
Closes #34