fix(vanilla): fix base path handling for vanilla push state
This fixes the vanilla PushState location service handling of `base` tags.
This change attempts to match the behavior of https://html.spec.whatwg.org/dev/semantics.html#the-base-element
If a base tag is '/base/index.html' and a state with a URL of '/foo' is activated, the URL will be '/base/foo'.
If the url exactly matches the base tag, it will route to the state matching '/'.

This also addresses a longstanding bug where base tags which didn't end in a slash (such as '/base')
had their last character stripped (i.e., '/bas'), and therefore didn't work.
Now base tags like that should work as described above.

Closes https://github.com/ui-router/core/issues/54
Closes https://github.com/angular-ui/ui-router/issues/2357